# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Gig Connect India API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WorkerRegistrationInputWorkType = typeof WorkerRegistrationInputWorkType[keyof typeof WorkerRegistrationInputWorkType];


export const WorkerRegistrationInputWorkType = {
  delivery: 'delivery',
  ride_sharing: 'ride_sharing',
  domestic: 'domestic',
  construction: 'construction',
  freelance: 'freelance',
  other: 'other',
} as const;

export interface WorkerRegistrationInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /**
     * @minLength 10
     * @maxLength 15
     */
  phone: string;
  /** @maxLength 254 */
  email?: string;
  /** @maxLength 100 */
  city: string;
  /** @maxLength 100 */
  state: string;
  workType: WorkerRegistrationInputWorkType;
  /**
     * Platform they work on (e.g. Swiggy, Ola, Urban Company)
     * @maxLength 100
     */
  platform?: string;
  /** @maxLength 500 */
  message?: string;
}

export interface Worker {
  id: string;
  name: string;
  phone: string;
  email?: string;
  city: string;
  state: string;
  workType: string;
  platform?: string;
  message?: string;
  joinedAt: string;
}

export interface ContactInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /** @maxLength 15 */
  phone?: string;
  /** @maxLength 200 */
  subject: string;
  /**
     * @minLength 10
     * @maxLength 1000
     */
  message: string;
}

export interface ContactMessage {
  id: string;
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
  submittedAt: string;
}

export interface CommunityStats {
  workersConnected: number;
  cities: number;
  welfarePartners: number;
  eventsOrganized: number;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export type ListWorkers200 = {
  workers: Worker[];
  total: number;
};

export type ListContactMessages200 = {
  messages: ContactMessage[];
  total: number;
};