# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gig Connect India API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers/register:
    post:
      operationId: registerWorker
      tags: [workers]
      summary: Register a new gig worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerRegistrationInput"
      responses:
        "201":
          description: Worker registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Worker already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    WorkerRegistrationInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        phone:
          type: string
          minLength: 10
          maxLength: 15
        email:
          type: string
          maxLength: 254
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 100
        workType:
          type: string
          enum: [delivery, ride_sharing, domestic, construction, freelance, other]
        platform:
          type: string
          maxLength: 100
          description: Platform they work on (e.g. Swiggy, Ola, Urban Company)
        message:
          type: string
          maxLength: 500
      required: [name, phone, city, state, workType]
    Worker:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        city:
          type: string
        state:
          type: string
        workType:
          type: string
        platform:
          type: string
        message:
          type: string
        joinedAt:
          type: string
          format: date-time
      required: [id, name, phone, city, state, workType, joinedAt]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]