# lib/api-zod/src/generated/types/worker.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Gig Connect India API
 * OpenAPI spec version: 0.1.0
 */

export interface Worker {
  id: string;
  name: string;
  phone: string;
  email?: string;
  city: string;
  state: string;
  workType: string;
  platform?: string;
  message?: string;
  joinedAt: Date;
}