# lib/api-zod/src/generated/types/workerRegistrationInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Gig Connect India API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkerRegistrationInputWorkType } from './workerRegistrationInputWorkType';

export interface WorkerRegistrationInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /**
     * @minLength 10
     * @maxLength 15
     */
  phone: string;
  /** @maxLength 254 */
  email?: string;
  /** @maxLength 100 */
  city: string;
  /** @maxLength 100 */
  state: string;
  workType: WorkerRegistrationInputWorkType;
  /**
     * Platform they work on (e.g. Swiggy, Ola, Urban Company)
     * @maxLength 100
     */
  platform?: string;
  /** @maxLength 500 */
  message?: string;
}